+++ src/components/grid/GridContext.ts
@@ -1,0 +1,9 @@
+import * as React from 'react'
+
+type Context = {
+  spacing: number | string,
+}
+
+export default React.createContext<Context>({
+  spacing: 0,
+})